docker from docker test pipeline
custom pipeline w/ docker client inside the docker container installation
and docker socket mount for inception.

the example use case for this is to have docker client for pipelines in
a travis build testing against php 5.3 for which travis would need the
precise image which has no docker, so php 5.3 comes via a docker image in
which by this feature pipelines *could perhaps* be tested.

additionally:

- phpunit testing across multiple versions incl. PHP 5.3

- alpine / ulibc support (no GLOB_BRACE)

- PharBuild\Builder is for PHP 5.4+ only

- fix test: even root can not unlink directories