fix for docker image name validation
regular expression for name-components was a name-component only. fix
here is to make it a product of name-components.

additionally reducing the regex for host-name to allowed characters
according to wikipedia [1] regarding allowed characters but still not
label arrangement and their length.

additionally corrected the documentation file and the comment within as
this flaw somewhat was expected and with this fix it is quite a bit more
appropriate.

refs:

- [1] https://en.wikipedia.org/wiki/Hostname#Restrictions_on_valid_hostnames