introduce file image types
an image is not always represented by 'a string' (the name) but can also
be 'a section'.

adding a concrete class for the bitbucket pipelines file model.

the images objects are lazy loaded and when converted to string retain
the old behavior.

also add image-name, another new file sub-type as parsing into
components is likely to happen on the image name string for private
repositories and the docker login command. when converted to string it
retains the old behavior as well.