add argument and option iterators
a lot of argument parsing is done by iterating over arguments and
more specifically options.

an argument iterator implements iteration over arguments. the getNext()
method is for some common look-ahead jobs.

an option iterator that encapsulates this is of use. it aggregates
the arguments iterator with two additional methods hasArgument()
and getArgument() so common options jobs are supported.

in addition to the iterator, a new filter iterator is directly looking
for one or more specific options.

this change removes the parts extracted from args.

fix of a bug removing only the last option when obtaining it's value,
fix is to remove the first option and value and leave the others as is.