add --no-keep option
to never keep containers, even on error. useful when failure/error running
a pipeline step is expected to fail. for example the composer ci script
which does a php lint check which can fail but there is no use to keep
the container for later.

additionally add that bewritten composer ci script to run all local
integration checks and tests, useful for finishing a local development
iteration and before pushing:

    $ composer ci