build phar api version
shown in the info about the phar file.

there are one-byte differences in the phar build depending on the php
version used, this might give a hint where the byte difference result
from.

note: looking into the future: the same api version is used, the
      differences are more likely to the change of the working tree when
      the phar is build.

additionally a minor fix in the travis build as update after require w/
update dependencies is not necessary.