introduce properties
simple value object of an array-near named entries based properties type.

intended for use in the image object as it has additional properties
likely to change over time and the operations with the data is quite
generic.

functionality is extracted in part from the image type which already
aggregates it and additionally property checking both required and
optional.