private docker hub support
if an image is used with docker up login credentials, checks if auth for
docker hub already exists, and if not, logs into docker hub first.

this change allows to use docker hub private repositories with support of
local docker hub authentication (~/.docker/config.json) to work offline
w/ copies of private images that are already available offline.

additionally environment variable name only settings in the image section
are resolved against the runner environment. use --env/ --env-file
settings to set/import environment variable(s).

note: not logging into docker hub is prone to errors in case a different
      docker hub configuration is shared via environment variables /
      pipelines file while the local system is already authenticated for a
      different account.

note: yet unknown is the handling with docker-machine as it is untested.

Signed-off-by: Tom Klingenberg <tklingenberg@lastflood.net>