fix deploy copy permission denied noise
previously when copying files into the container for deploy copy mode
and a file was not accessible (permission denied error), the docker cp
command had potential to spill out not only the error on the file but
also on any other file following. depending on the number of paths to
copy into the container, this flooded the shells error output to such a
large extend, that the original error was hidden due to the limit of the
scroll back buffer and the shear amount of noise.

fix is to tar the files and pipe the tar to the docker cp command.

the error output is now from the tar command alone, the copy operation
is not interrupted any longer and runnint the pipeline continues:

    +++ copying files into container...
    tar: ./icons/symbol.gif: Cannot open: Permission denied
    tar: Exiting with failure status due to previous errors

the change has potential to run a container on an incomplete deployment
w/o a non-zero exit status of the pipelines utility itself. but using
pipelines in development has shown that it is better to design for
failure at this point and ignore a copy failure as a clean checkout
would not have contained the file which caused the error.

for strict monitoring of pipelines errors can be found in standard error
for the deploy copy operation.