add .env / .env.dist file support
use .env.dist and .env files from working directory if any of these
files exist.

first .env.dist is added, then .env. this order allows to overwrite
(default) exports or sets with a .env file which is intended to be
ignored in version control.

this is as if variables are configured for the respository or team on
bitbucket.