fix build directory owner and attributes
containers with a user different to root were not compatible as the root
folder of the docker cp operation was owned by root instead of the default
container user.

fix is here to put the directory basename inside the tar by making use
of a symbolic link which is de-referenced.