streamline argument errors
for one that is the style of error messages (which now shows an option
name always prefixed with a single or double hyphen / minus sign(s)) and
for the other it is to show these errors as application errors on the
utility level (like with other non-zero status, stderr and prefixed with
utility name).

additionally minor fix in case of unknown option provided to give an
argument error not just the non-zero exit status.