fix keeping containers
due to some bit mask confusion in recent change to the default "never keep
a container", not even on error, the --error-keep flag had no grip.

fix is to add the flag (instead of removing it) when --error-keep is
given.

additionally prevent the duplicate output of the exit status in case of
keeping a container on error.