fix coverage checker script precision
when the two-digit round (half up) value was exactly like the expected
value, it still was lower so the message given on not matching was not
looking good:

    Code coverage is 21.72%, which is below the accepted 21.72%

fix is to always round to two digits for percentage values.

this will also make these special cases pass which most often should be
fine unless we learn that two digits is just two little for percentage
values to catch the edge cases.