fix duplicate error code output
previously thought of as a good idea (tm) turned out to just introduce
duplicate output which somewhat is a very bad looking thing and does not
at all help w/ usability.

fix is to go back to the original output which worked and to benefit from
revisiting diverse output points.