polish
switch self to $this - just b/c the rest of the code does. when reading
about changes for php 7.0 self is probably more native to the language,
but while in review w/ PSR-5 draft $this is fine as well and as the rest
of the code base makes use of it, this is it for the moment.

running inspections in phpstorm and then doing some changes. this includes
adding phpunit metadata for used mock methods. kudos to ea inspections.

additionally return typehint and check scripts w/ shellcheck and other
minor CS fixes.