reduce chunk size
the size of 1792 lead to a real-life 127 exit status while copying
artifact files, so reduced to a working variant as fast work around.

last time the size was changed in 5d58a59 in May 2018 from 2048 to
1792 for the same reason. turned out, reducing further down to 1512
did not work out well enough as it still was giving error:

     Artifact failure: 'foo' (127, 1512 paths, 137577 bytes)

switch to a MAX_ARG_STRLEN (131072) [1] string length limitation of the array
chunks.

Which is prevented by the 131072 max length setting in this change set.

[1]: https://www.in-ulm.de/~mascheck/various/argmax/
Reference: 5d58a59da5baf4fccf74ad2098794db126111d36