add php execution method to phar builder
for commands / utilities which are php based (like composer), it can be
wanted to execute these commands not with the shell's default php binary/
interpreter but with the version the phar build php script was called.

if so the the Builder::phpExec() method is a drop-in replacement for the
Builder::exec() call which assumes the utility in the command is a shell
resolvable path to a php script which then is invoked via the build scripts
php binary.

this allows to make a phar build script more portable and to test the
phar build against different php versions more easily (which requires the
build script to make use of this new method).

NOTE: this new method depends on a shell feature: the "which" utility.

additionally change some double quotes to single quotes as EA Inspections
complained, a bad habit maybe.