fix getting the version from git
previously the git version was returned from the git command w/o taking
the field separation into account.

fix is to quote the actual value so that changes to IFS or similar in
shells different to bash or zsh.

NOTE: test was to set IFS to a single dot "." which produced many more
      interesting results. take care when doing this at home.

reference: https://unix.stackexchange.com/a/68748