ease build path segment pattern use
previously the double star pattern did not match zero path segments which
required it to be written as an option of a brace.

change is to make that workaround not necessary any longer by moving it
towards expanding the pattern.

this should come closer to existing pattern matchers, e.g. patterns in
the .gitignore files.