re-use kept containers
while developing with a pipeline it came to attention that the --keep
or --keep-on-error switches did not provide any benefit when running the
pipeline again as the previous container was always zapped.

change is if any of the keep flags are set and a container with the name
for a step already exists, that it is reused.

and more precisely set the default flags in tests.

additionally check phpunit version before re-installing it. it works but
is not useful unless files aren't copied into vendor / composer.lock any
longer for deploy copy on kept containers.