execute script as script
previously each command line of a script has been passed as a single exec
call each into the container.

change is to turn it into a script first which is similar to how it is
said to be done in the cloud [1].

this allows shell variable handling in the script itself.

[1] https://community.atlassian.com/t5/Bitbucket-questions/In-Bitbucket-Pipelines-why-doesn-t-get-the-expected-exit-code/qaq-p/612957