add binary docker packages
using the docker client in any container "as a service" could be done by
injecting the static docker binary into the running container while
having the socket mounted so that docker can be used from inside the
running container even if the container image itself has not docker
client. this is practical as it provides docker as a service which
aligns with the YAML notation.

however to obtain the static docker binary requires going online and
fetching the distribution package and untar the docker binary from it.

this is done by the binary unpackager, it un-packages binary docker
client. the static binary is aimed for so that this is a single file
install for having docker client inside the pipeline environment.

this very first implementation offers a simple package format and
download as well as un-tar verfification with SHA256 hashes.

once obtained and the pipelines data directory is not wiped, the download
is not necessary any longer.

packages can be written as YAML files, the in-memory representation is
just an array in this implementation. comments on the entries can be
found in the test-package YAML file.

file_get_contents to read .yml files from within phar files.