add environment doc
Atlassian Bitbucket Cloud Pipelines Plugin added a couple of additional
environment variables over the years since project start. Most of them
can be handled by the pipelines utility albeit not directly supported.

Some not, which do highlight missing (unsupported) features.

Documenting a list of all (current) variables is helpful to get an
overview about how the super-set of environment parameters is dealt with
(or not) and also on how to deal with environment variables on a project
level the pipelines utility was never to be considered compatible with
at all in a practicable manner.