improve runner directories
albeit XDG is for a standard, the use in pipelines is always bound to the
project itself - and it's utility name. Therefore there is no generic XDG
base directory class yet, just one that adheres to the specification in
context of the project.

FIXME: there is some FIXME still to do but the first feature that relies
       on the newer XDGBDS should be done first.