improve pipeline script visibility
previously in f3ef6ff execution of the pipeline script as an actual
script has been added.

however that implementation did make use of temporary files to first
store and then pass into the pipeline container for execution.

this lacks transparency of the pipelines utility which offers the `-v`
(verbose) option to show _every_ command that is getting executed when
a pipeline is run.

the verbose output only showed the temporary pathname which is a) not
very readable and b) cumbersome to find out what the actual script is
and c) the temporary files had to be kept so that the user could review.

improvement here is to switch from creating a temporary file to here-
document. the script becomes directly readable and there is no need to
create any temporary files explicitly.

to support the changes a new test-driver has been added:

    $ tests/shell/run.sh test-polish.sh 1

a spurious new-line has been introduced when switching to script
execution back then.

while running over and over again right now, it seems misplaced and is
gone now.

Ref: f3ef6ff1ee63edd86097db5c0a635b8bfbcc93f4