fix builder ** limit expansion
when building the phar file with the src/**/*.php pattern w/ more
subdirectories now (docker binary repository below runner) it became
visible that the brace expansion for the ** pattern did not work.

somehow moving all opening braces after each other on the beginning - not
at the end as previous - did the trick.

however it became also clear that the expansion routine Glob::expandBrace
is broken. Perhaps through self::expandBraceInnerMatch. To address the bad
test situation, add a regression test to show the buggy behavior and
document the bug in the read-me.