add static factory method to runner
the logic to create a runner is intern / extern due to creation of objects
inside the runner ctor.

extracting a static factory method allows to pull out that creation logic
from the ctor allowing easier changes and more overview which turns out
it's not in production use.

keeping the name w/ the "Ex" suffix for the moment, enforcing use in the
utility App implementation as to target there.

is refactoring of the runner.