improve lib fs
add portable file-name support:

thanks to the much appreciated publishing work of David A. Wheeler, this
one was easy to implement as needed to make things portable.

LibFs::isPortableFilename is to check if a file-name is portable.

allow phar:// scheme while normalizing a path:

phar based paths are not overly well supported by lib fs albeit phar
format is part of the project build target.

also a fix for the strspn part which worked on $path instead of $buffer
and leaving some notes on php stream wrapper support.

add/extract lib-tmp:

library for temporary files and directories related functions.