fix exec tester on not passed tests
previously when a test did not pass, the exec-tester did do additional
assertions which then could also fail.

in case of failure, this tricked phpunit into only reporting the assertion
problems of the exec-tester hiding the actual test failure as a test
can only be non-successful for a single reason.

fix is to check the state of the test-case first and if not successful
to skip the internal assertions which are triggered due to destruction
after the test has finished.

this fix also made a couple of tests visible which were not successful as
they had no assertions (test "fail" being risky). for these tests add at
least one assertion.

additionally bind the exec-tester to the pipelines test-case sub-type
instead of the phpunit test-case to narrow its use-case.