beautify error output on injection failure of docker-client
put the error text onto a line of it's own when viewing stdout and stderr
at once as common in shell.

this makes it more clear that there was a pipelines internal problem
injecting the docker client, not a system issue. for example providing a
wrong argument can trigger these kind of errors.

found while testing on the command-line.