support resolving local binaries
allows to use binaries available on disk (pipeline inside pipelines not
yet supported).

one caveat: when using a concrete docker client binary in the local
directory with a relative path, the path to the local docker client
executable is only taken if the argument contains a slash which allows to
create binaries having the same name a package:

    $ bin/pipelines --docker-client ./docker-42.42.1-binsh-test-stub
                                    `^``````````````````````````````
                                     `-- with slash

is using the binary named docker-42.42.1-binsh-test-stub in the working
directory in difference to:

    $ bin/pipelines --docker-client docker-42.42.1-binsh-test-stub
                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                                       `-- without slash

which uses the built-in package named "docker-42.42.1-binsh-test-stub".