scrutinize and improve
review a couple of places that scrutinizer highlights.

it's known the bootstrap fails if STDERR / php://stderr is not available.
this *should* never happen, if it happens, one bail out could be to switch
to echo.

the phpstorm meta annotations got an update on the annotations.

some little improvements on type-hinting (per doc-block).

in file, found an issue where a integers could slip in for pattern, string
cast for a fix.