add busybox image for bitbucket pipelines
while testing pipelines on bitbucket with the official ducker hub busybox
image, the atlassian bitbucket cloud pipeline plugin ran into an error
while setting up the build [REPORT]:

    /bin/sh: /usr/bin/mkfifo: not found

to work around this on atlassian bitbucket a busybox image that is just
the busybox image only modified w/ a symbolic link to prevent the error
has been created and uploaded.

in this change set is the script to build and push the image as well making
use of it in the projects' pipelines.

this should make pass the build again on atlassian bitbucket cloud.

it might be that w/ future changes on atlassian bitbucket cloud the
original image is supported (again).

REPORT: https://community.atlassian.com/t5/Bitbucket-questions/Pipeline-Build-Setup-bin-sh-usr-bin-mkfifo-not-found/qaq-p/1253290