fix mount after copy in pipeline inside pipelines
when running a pipeline inside a pipeline that use deploy mode copy,
--deploy copy (current default), and that pipeline is using deploy mode
mount, --deploy mount, then project was empty as the original host
path to mount could not be identified and it remained /app which was
not mountable resulting in an empty directory at /app in the inner-most
pipeline.

first of all there is the issue here of not seeing any error that the
mounting "failed". which is b/c running the pipeline container is
successful, especially with mounting a non-existing directory outside of
any container as docker creates it then.

as this is not an error situation - also if the host path consists of
multiple recursive directories - re-mounting the same directory now emits
an error and brings the pipelines utility to an early halt.

NOTE: as docker is that invasive, an additional check should be added
      to prevent mounting any path not from a safe location, e.g. from
      a logically resolved ^/home/[^/\0]*/[^./\0][^/\0]*/ (realpath not
      possible inside a container) and /tmp/... for testing (where it
      must exists, for integration testing, expect the project be under
      home). For build travis and scrutinizer are within $HOME as well,
      bitbucket pipelines is not so it must allow BITBUCKET_CLONE_DIR

          /opt/atlassian/pipelines/agent/build

      under circumstanes, too.

adressing the device-dir as mount point with this first sanity check
should help in preventing the issue and report errors early.

to enable pipelines to mount inside a pipeline with --deploy copy, the
original project-directory is set in the runner environment in the topmost
pipeline run as PIPELINES_PROJECT_PATH. this parameter then is used if
docker inspect can not reveal the device-directory as there is no mount
point inside a pipeline of --deploy copy.

NOTE: when running a pipeline, this variable can be changed. this will
      allow to inject a device directory that perhaps should not be
      mounted into another container. this makes the first note that
      important. previously this was not possible.

the NOTEs must be addressed in a future change.