add application options object
pipelines has scattered some non-shell parameters which are used
internally only and most often hard-encoded so far.

adding an application options object that keeps the parameters at a central
place (bookkeeping) to be injected via the run opts or even differently
should allow to better control run-time behaviour also within tests.

the get method is expected to be called with strings as names so that only
the option name is hard-encoded but not the value of it.

add phpstorm completion for fun and profit.