add --docker-client-pkgs to list built-in packages
a listing of all built-in docker client packages was not available
previously.

adding the implementation and binding to a command-line argument so that
it can be tested with the phar file.

additionally validate parameters on args parsing and present extensive
error information to the user for --docker-client which includes
showing a list of available packages in the error message.