mount docker client instead of copy
previously the docker client integration is by copying it into the
container (copy the binary into the running container).

change is to mount it instead.

this is to better support pipelines inside pipelines (pip) so that the
same docker client can be re-used without downloading the package into
the parent container.

this re-downloading prevents offline usage, which is addressed by this
change.

a downside of the mount is, that the docker client package can not be
changed in pip as the original host docker client is re-used. However this
scenario is less likely currently.

adds support for read-only mounts, the docker client is mounted read-only.

integration tests / composer ci