add --step/s option
in normal operation all steps of a pipeline are executed.

in case of failure or extending an existing pipeline, previously it was
not possible to control which steps of a pipeline are to be executed.

the --step (and --steps) option control the number of steps to be
exucuted. when in use, only these steps are run by pipelines.