fix phar build since 0.0.25
(on travis,) running the phpunit tests before the phar build taints a
timestamp on the repository file tests/data/package/docker-test-stub
making git diff-index --quiet HEAD -- return non-zero which is adding the
"+" (plus-sign) to the version resulting in pipelines phar version with
that adding plus-sign albeit the repository version build from is fine as
that file did not change (content) nor is it added to the phar.

this leaves the packages offered via github (which is also for phive)
not easily reproducible as the wrong version identifier is used.

fix is to make the build obtain the version with the --dirty flag as it is
immune to such changes since git 1.7.7.

alternatively it would be possible to call "git status" or refreshing the
git index first to levitate the problem, but as git is fixed for the
--dirty flag since years, this seems preferable, especially in context of
the travis build.