fully ignore platform reqs in phar build
previously only on the re-install (with dev dependencies) the platform
requirements were ignored (composer install).

as it looks like an oversight and it makes no sense to only do so in the
second install command, also do so on the first install when the develop-
ment dependencies are removed for the build.

ignoring platform dependencies during the build makes sense to use a less
(or little) configured php runtime as for building (only) the platform
dependencies are not (or even should not be) required. at least this should
be more portable for the build (e.g. ext-json is not required for the
build).