composer version in phar build
composer is an integral part of the phar build due to the autoloader in
use.

but also depending on composer version, the build binary output differs
(most likely due to different creation of files in vendor/composer folder).

starting with this change-set the composer version is pinned in:

    lib/pipelines/composer-install.sh

therefore giving more info about the build-system is useful to learn about
in which boundaries the phar file is created.

additionally pining the composer version in the build.