which composer / php binary
obtaining the path to the composer executable it not always possible with
the shell. e.g. `php -f $(which composer) -- ...` is not compatible with
travis (or wrappers similar to the one one on travis) as $(which composer)
does not return the path of the composer (phar) executable.

using it as the php script to run, it only outputs the executables content.

add a composer script "which" that is able to return it's own location w/
the COMPOSER_BINARY environment parameter available in composer scripts.

e.g.:

    $ composer which 2>/dev/null
    /home/user/bin/composer.phar

the example shows it also effectively resolves symbolic links:

    $ ls -al "$(which composer)"
    lrwxrwxrwx 1 user user 1 Dez 17 12:34 /home/user/bin/composer -> ./composer.phar

another usage example:

    $ php -f "$(composer which 2>/dev/null)" -- ...

additionally add another script "which-php" for the php command in use:

    $ composer which-php
    > echo ${PHP_BINARY}
    /usr/bin/php7.4

or:

    $ php7.2 -f "$(composer which)" -- which-php
    > echo ${COMPOSER_BINARY}
    > echo ${PHP_BINARY}
    /usr/bin/php7.2

all example lines starting with ">" are on stderr.