fix build for phpunit 7.5+
somewhere between phpunit 6 and 7.5 the behavior of STDERR being available
has changed. while it was not available with phpunit 6, it is available
with phpunit 7.5.

as the work-around code was under coverage report and it does not kick
in with phpunit 7.5, the build failed.

excluding from coverage alleviates the problem and stabilizes for
phpunit 7.5+.