add script.exit-early configuration
of type bool.

on bitbucket pipelines set -e is the default behavior for pipeline scripts.

pipelines offers a more strict check after each command so that any side-
effects to rely on set -e are eradicated by checking the actual exit
status of a step scripts command (one command is a shell pipe from an
array entry in the step.script array).

this ensures to exist for the following conditions (not catched on
bitbucket pipelines):

flaw on use of `set +e` in step script:

    - step:
        image: alpine
        script:
          - set +e
          - mkdir -p /dev/null
          - echo "surrender"

flaw on use of `&&` in step script:

    - step:
        image: ktomk/pipelines:busybox
        script:
          - mkdir -p /dev/null && test
          - echo "surrender"

fix for strict exit early handling is to check for the exit status
explicitly after each command and exit if non-zero.