add support for BITBUCKET_STEP_RUN_NUMBER
by default BITBUCKET_STEP_RUN_NUMBER is "1", on bitbucket if a failed step
is run again, it is count up (e.g. "2") [1].

in pipelines it is inherited from the environment, defaults to "1" and
reset to "1" after each pipeline step to "1".

example:

    $ BITBUCKET_STEP_RUN_NUMBER=2 pipelines --step 2

run the second step as if it would be run as the first re-run after
failure.

[1]: https://community.atlassian.com/t5/Bitbucket-questions/About-the-BITBUCKET-STEP-RUN-NUMBER-Default-Variable/qaq-p/1319257