label containers
for all container types so far:

- step
- service

for now all labels start with "pipelines." in their name.

the standard labels are:

    pipelines.prefix   prefix in use. if not set then default "pipelines"
    pipelines.role     "step" for step-container, "service" for service-
                       container
    pipelines.project.name
                       name of project (directory base-name)
    pipelines.project.path
                       path to the project

labels are at time of last use. e.g. if the prefix or the project
changed for the same image, the label values are from that last container
run.

add labels builder collaborator to support this, especially service
containers are still with very little abstraction - if at all.