add step.clone-path configuration (2)
of type string.

mount point / destination of the project files withing a pipeline step
container.

previously the path to the project mount/copy/clone in the step-container
was hard-encoded as "/app". this configuration option is for those many
string places.

first part done in 75fca8f, some changes lost in translation added here.

ref: 75fca8fcb79796d30d370f26458cd52cc95e4c47