add docker abstraction layer
Defined in runner for its use, but first only used in cache-io.

From the runner side, also (but still unused there, too) inspected from
the cache-io. Let's see how to get another layer into there to get the more
low-level stuff ready for dependency inversion.

tests come two-folded, unit and integration. unit is completely mocked,
integration needs the docker client to be available in tests.

for starting containers, only those supporting /bin/sh with a prompt are
supported (which is current understanding of how it works for pipelines,
binding to Linux containers and adhering to posix).