migrate docker client install to http-cache
in f73257cd an example of how to migrate from the very early offline
workaround as been given incl. changing the (old)docker client install
script and the bitbucket-pipelines.yml file.

this change addresses parts of the pipelines yaml file that lost their
cache due to the changes for downloading the docker client slowing down
the build.

fix is to add the build-http-cache to docker client affected pipline steps.

ref: f73257cdd7509f6dd97070d298c30d9c33dbb91c