code fixes: type analysis
code-fixes being more strict on type annotations and types in general.
this is the result of reviewing type check errors for the psalm-ification
of pipelines.

for runnable interface added one specifically for
returning integer (more close to status 0-255).

templating for args get-option-argument as the default value
influences the return type; also moving from bool to int, as no
bool was in use albeit technically possible.

one real fix where the code was actually wrong in pipeline
filter-map-list.

another real fix where comparing against a wrong return value
of preg_replace (never returns false on error but null).

make types more specific with iterables/traversables. and it actually is
some psalm template fun for iterators reading helps:

https://psalm.dev/docs/annotating_code/templated_annotations/

no implicit string conversions, changed to string casts. similar
for where strings were used as integers.

for the phpunit phpt workaround, suppress for psalm. this might
be wrong as the safe-guard checks might not apply (any longer/at
all), but as feeling weak with that part, opting for suppression.

null-able types both in and out pose some complexity issues, fix strategy
is to give optional parameters true defaults if possible to remove the
null-able on input directly.

if not possible to throw when going out and still null.

otherwise use casting on next use.

places that caught attention and are questioned by the design:

for more type specific getters (as previously in args and now
showing for the options):

- getting from the options could just throw when an option does
  not exist. perhaps on more type specific getters, like getting
  a string option.
- with such type specific getters has been experimented with the
  args class, turned out this was only differentiating for the
  type of the default value input parameter, which for the less
  typed getter, psalm template types are in use now which might
  allow to remove the more typed getter later on.
- first fixup for options, reducing on the runopts to get strings
  only and throw if n/a. looks like the null was superfluous.

first assertion by @pslam-assert-if-true, specifying string/int options
(aka enums) is not possible though. established the constants anyway,
might hurt some day, now is in the file component runner depends on
anyway.

psalm also reveals places where psalm can't detect the type
properly, esp. when the code branches based on type. e.g. in
the script runner for the different kind of step entries (line
vs. pipe). here the branching was already correct but psalm needs
additional @var type-hints and/or castings to get it right. as
these places are merely locally, this adds some additional
documentation in the code which should remove some mental over-
head as well when going back there-in.

psalm needs info about matches from preg_match or it's me not
seeing it might still be null?! there might be some options to control
this more properly for the return type, but they make this more strict
so would be more detail there. it's perhaps that psalm might benefit
from better stubs for internal. tbd.

improve type annotations of the base test-case.