update php-cs-fixer and add custom cs-fixers lib
thanks to @kubawerlos there are some php-cs-fixers that match my needs and
he was so kind to point me to them.

this requires keeping php-cs-fixer on the latest stable release, therefore
updating it:

    $ composer update friendsofphp/php-cs-fixer

and then installing the customer fixers:

    $ composer require --dev kubawerlos/php-cs-fixer-custom-fixers

this commit prepares running next cs checks with more fixers, including
one for having missing spaces after foreach which scrutinizer rightfully
reported which surprised me as i was under the impression this should
have been taken care of automatically so far.