fix broken artifacts
since 0.0.43 artifacts were broken.

adding the step.clone-path option in 75fca8f switched to existing property
of the clone path in ArtifactSource but ArtifactSource was created with
the wrong path when in use: instead of the in-container path the project
directory was provided.

Fix is to instantiate ArtifactSource with the correct directory.

Additionally add a small shell-test to confirm artifacts working testing
for the functionality.

ref: 75fca8fcb79796d30d370f26458cd52cc95e4c47