fix readable local stream detection
just now composer ci failed on test-case LibFsStreamTest::testIsReadable
for false for http://ktomk.github.io/ (http://...).

there were no changes in the code-base regarding this (?).

fix is to add a check for stream to be local. this is what is wanted
in any case.

cause is unknown, but not the php version (see travis logs, just all
php versions had this test failing).