add html-docs (mkdocs + material)
html documentation from project markdown files by the nice combo of
Mkdocs and Material for MkDocs (mkdocs, mkdocs-material).

html, multiple files. includes javascript search. foundation of the github
pages based [Project-Homepage], baseline for html documentation available
with the project to easily browse with a www/mosaic client from local
via file:// scheme, no implicit WAN.

caveat: mkdocs expects all documentation files within a folder, [docs_dir].
as pipelines has the markdown files already in a different file-layout (and
mkdocs can not easily use the projects root), a Makefile is in use to
compile everything in the expected format (mainly making use of symbolic
links as imports, see import.lst).

this is not that bad anyway as some part of the documentation is generated
from a project checkout / composer install like the licenses in use and
some more tweaks are in effect to handle offline-usage.

includes an example to migrate between mkdocs directory url layouts.

lxml is part of the mkdocs requirements (for pipelines, not default mkdocs)
and in use for the gh-pages build url migration.

[docs_dir]: https://www.mkdocs.org/user-guide/configuration/#docs_dir
[Project-Homepage]: https://ktomk.github.io/pipelines