add php project configuration script (ppconf.sh)
/bin/sh shell script w/ directive arguments for ci related configuration:

    usage: lib/scripts/ppconf.sh [directive [parameter ...] ...]

    Directives
        help                  show usage help
        self-test             show self/system info
        disable-tls           http only composer
        phpunit[-version] <ver>
                              set phpunit version, <ver> example: ^4
        downpatch[-phpunit]-tests
                              testsuite phpunit compatibility (phpunit < 8; php 7.4 open-up)

examples:

  ./lib/scripts/ppconf.sh help
  ./lib/scripts/ppconf.sh self-test
  ./lib/scripts/ppconf.sh disable-tls
  ./lib/scripts/ppconf.sh phpunit-version ^4
  ./lib/scripts/ppconf.sh downpatch-phpunit-tests
  ./lib/scripts/ppconf.sh xdebug3

the phpunit test-suite patching is what it was previously but in both
directions (patching up and down):

    $ ./tests/shell/run.sh test-php74.sh 2

adding -2 for the inverse there-in, too.

there was a php version break-point with php 7.4 which requires phpunit 8+
and phpunit 8 introduced syntax changes on the PHP level that required the
backwards compatible phpunit test-suite in pipelines to be patched to
match phpunits PHP syntax requirements (e.g. return type :void
declarations for test-case pattern methods, inheritance is w/ the public
interface) as pipelines supports PHP 5.3.x - 8.x with a single phpunit
test-suite (preferring expression).

previously the solution was to patch the old code and lift it up for more
recent PHP versions. better is to shift current and make it backwards
compatible. this is how proceedings will go. took me a moment, but php 8
shows that there will be more language changes on the syntax level and
then again (and again, etc).

it is also much more comfortable this way round and also for streamlining
processes across bare-metal dev and all containerized / virtualized
integration automations.

next change is to have phpunit 8+ compatible code (php 7.2+) in the project
and inverse the patching to make the existing code backwards compatible for
integration with PHP versions below 7.2.

with/has the benefit to run the test-suite on checkout on any php 7.2+
systems (currently). these are also effectively the systems supported by
xdebug 3.0 (xdebug is used in the project setup for code-coverage as it
is best supported across versions) and one of the further expected changes
are to support xdebug 3.0.

choosing /bin/sh run directly within the barest containers (e.g. busybox
based). better out-of-the-box experience.

next to porting the patch from ./tests/shell/test-php74.sh (and extending
it), the routine to handle phpunit installation via composer is ported
from the projects bitbucket-pipelines.yml file as well to centralize these
project related routines.

the exemplary use-case is the custom/unit-tests pipeline which runs on
php 7.0 and did not had full composer support before. when switching the
default project phpunit version from ^6.5 to ^8, it is required for
all pipelines below php 7.2 to downpatch the test-suite _and_ reconfigure
the phpunit version.

a test are the php 5.3 tests and the linting pipelines as the cover 5.3
and 7.4.

NOTE: downpatch-phpunit-tests is idempotent and is behaving when running
      on a non-patched phpunit test-suite. changes to the test-suite and
      change of the project phpunit base-line version are still pending.

next to phpunit versions there is also xdebug versions. they are
handsome in general, the breaking point is with the .phpt tests that are
run by phpunit. as xdebug 3 changes some of the ini configuration
settings an additional patching is necessary as otherwise the blackbox
.phpt tests fail with different output (xdebug 2 and 3 ini settings are
generally incompatible to each other).